Remove dependencies on analyzer libraries leaking through to runtime use
There currently is only one assembly that is duplicated to act as Roslyn
analyzer and runtime code for the "Open" attribute. This results in
projects in a non-analyzer context also needing to load the analyzer
code's dependencies (such as Microsoft.CodeAnalysis.CSharp.Workspaces),
which in turn causes problems for Unity (3D) projects because Unity
separates assemblies intended as Roslyn analyzers and runtime code;
as a result, Unity tried to also load the other analyzer dependencies
for runtime application code, which it doesn't have by default because
it has a more limited .NET Standard 2.1 runtime (in Unity 2021.3).

Since the runtime code (just the attribute) doesn't actually need these
dependencies, split off a runtime project and assembly to fix the issue.

Fixes #4